[upstream] Kinematic perf (erincatto/box2d#970)
Kinematic perf (erincatto/box2d#970)

- b2Shape_SetMaterial -> b2Shape_SetUserMaterial
- Reworked surface material access on chain shapes
- Removed mouse joint. Functionality is now covered by the motor joint.
- Added b2DynamicTree_QueryAll to support debug draw for shapes that
don't collide
- The user material id is now a 64 bits
- Fixed crash when samples app is minimized
- Added kinematic washer sample
- Fix erincatto/box2d#967
- #968 - draw shapes with 0 category bits and draw sleeping contacts
- Kinematic bodies no longer enter graph coloring
- The joint and contact solvers no longer write the state of kinematic
bodies
- Large kinematic bodies touching many dynamic bodies now has better
multithreaded scaling
- Joints between two non-dynamic bodies no longer simulate
- Minor island splitting optimization suggested by Dirk, use the island
id instead of a bool mark
- Joints now have a constraint damping ratio of 2.0 instead of 0.
Intended to reduce overshoot when limits are violated.
- Restored island debug draw

+++ src/Box2D.NET/B2Delegates.cs
@@ -57,13 +57,13 @@ namespace Box2D.NET
     /// from a worker thread.
     /// @warning This function should not attempt to modify Box2D state or user application state.
     /// @ingroup world
-    public delegate float b2FrictionCallback(float frictionA, int userMaterialIdA, float frictionB, int userMaterialIdB);
+    public delegate float b2FrictionCallback(float frictionA, ulong userMaterialIdA, float frictionB, ulong userMaterialIdB);
 
     /// Optional restitution mixing callback. This intentionally provides no context objects because this is called
     /// from a worker thread.
     /// @warning This function should not attempt to modify Box2D state or user application state.
     /// @ingroup world
-    public delegate float b2RestitutionCallback(float restitutionA, int userMaterialIdA, float restitutionB, int userMaterialIdB);
+    public delegate float b2RestitutionCallback(float restitutionA, ulong userMaterialIdA, float restitutionB, ulong userMaterialIdB);
 
     /// Prototype for a contact filter callback.
     /// This is called when a contact pair is considered for collision. This allows you to
